% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agenzie_provincia.R
\docType{data}
\name{agenzie_provincia}
\alias{agenzie_provincia}
\title{Agenzie e Subagenti per Provincia Italiana}
\format{
A data.table of 110 rows and 3 columns
\describe{
\item{provincia}{Character values indicating province ID}
\item{agenzie}{Numeric values containing the number of agencies in the given province}
\item{subagenti}{Numeric values containing the number of subagents in the given province}
}
}
\source{
Indagine: Agent Satisfaction
}
\usage{
agenzie_provincia
}
\description{
A table containing number of insurance agencies and subagents by italian province.
}
\examples{
head(agenzie_provincia)
}
\author{
Fabio Sgobio, Carlo Cardinali, Alejandro Abraham
}
\keyword{datasets}
