% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bareme.R
\docType{data}
\name{bareme}
\alias{bareme}
\title{Bareme Table}
\format{
A data.table of 324 rows and 3 columns
\describe{
\item{car_a}{interger-factor indicating Car A role in the incident}
\item{car_b}{interger factor indicating Car B role in the incident}
\item{outcome}{alpha factor indicating the outcome}
...
}
}
\source{
ISTAT
}
\usage{
bareme
}
\description{
A table insurance companies shorthand codes from 1-17 and all their combinations with the responsability outcome.
}
\author{
Fabio Sgobio, Alejandro Abraham
}
\keyword{datasets}
