% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
clean_names(dat, ...)
}
\arguments{
\item{dat}{is a data.table.}

\item{...}{other arguments coming from the janitor packages.}
}
\value{
data.table .
}
\description{
Snames case colnames.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
