% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codice_ateco.R
\docType{data}
\name{codice_ateco}
\alias{codice_ateco}
\title{Codice Ateco 2021}
\format{
A data.table of 3145 rows and 4 columns
\describe{
\item{cod_ateco}{Alpha-Numeric values giving the resulting Ateco code, ID}
\item{descrizione}{Character values giving the description, ID}
...
}
}
\source{
ISTAT
}
\usage{
codice_ateco
}
\description{
A long-format table containing all Ateco codes, their description and two Level 1 columns as shorthand.
}
\author{
Fabio Sgobio, Alejandro Abraham
}
\keyword{datasets}
