% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comuni_short.R
\docType{data}
\name{comuni_short}
\alias{comuni_short}
\title{Decodifica Comuni Short}
\format{
A data.table of 8258 rows and 5 columns
\describe{
...
}
}
\source{
ISTAT
}
\usage{
comuni_short
}
\description{
A table containing a lookup table ready for comune to area geografica.
}
\author{
Fabio Sgobio
}
\keyword{datasets}
