% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme.R
\name{create_readme}
\alias{create_readme}
\title{Create README.md files}
\usage{
create_readme(
  template = c("Overview", "Documents", "Code", "DataRaw", "DataProcessed",
    "DataOutput", "Figures", "Deliverables", "Reports"),
  path = getwd()
)
}
\arguments{
\item{template}{Which README.md files should be created? Default is all. Partial
matching does work.}

\item{path}{Where should they be created? Default is the working directory.}
}
\value{
This function creates the desired README files. It will not overwrite the file
however if it does not exists. It does not return anything.
}
\description{
This function creates README.md in the Innovation Team templates for the standard
file structure
}
\keyword{README}
\keyword{README.md}
