% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodifica_comuni.R
\docType{data}
\name{decodifica_comuni}
\alias{decodifica_comuni}
\title{Decodifica Comuni}
\format{
A data.table of 9081 rows and 20 columns
\describe{
...
}
}
\source{
ISTAT
}
\usage{
decodifica_comuni
}
\description{
A table containing from Comune to Region in diverse formats to convert from Old ID's to New ID's.
}
\author{
Fabio Sgobio
}
\keyword{datasets}
