% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_years.R
\name{diff_years}
\alias{diff_years}
\title{Years Difference}
\usage{
diff_years(base_date, tap_date)
}
\arguments{
\item{DT}{a data.table to obtain column names from.}
}
\value{
a one row data.table keeping original names.
}
\description{
Calculates the times passed between dates, e.g. how many years occured between 2020 and 2021 = 1.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
