% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_png.R
\name{export_png}
\alias{export_png}
\title{Export PNG with Logo banner}
\usage{
export_png(path_destination = NULL, widthIT = 10, heightIT = 6, dpiIT = 320)
}
\arguments{
\item{path_destination}{relative PATH where to save the plot}

\item{widthIT}{PNG width}

\item{heightIT}{PNG Height image}

\item{dpiIT}{PNG print quality}
}
\value{
A print/publication-ready PNG file with logo banner
}
\description{
Contains the parameters and colour for a standarized plot theme.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
