% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_italia.R
\docType{data}
\name{geo_italia}
\alias{geo_italia}
\title{Italian Geographic Levels}
\format{
A data.table of 7904 rows and 5 columns
\describe{
\item{nuts1}{Character values indicating Area ID}
\item{nuts2}{Character values indicating Region ID}
\item{nuts3}{Character values indicating Province ID}
\item{comune}{Character values indicating Comune ID}
}
}
\source{
EUROSTAT
}
\usage{
geo_italia
}
\description{
A table containing from NUTS1, 2, 3 and Municipalitie level and Comune Code for Italy.
}
\author{
Fabio Sgobio, Alejandro Abraham
}
\keyword{datasets}
