% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{\%contain\%}
\alias{\%contain\%}
\title{\%contains\% Operator}
\usage{
strings \%contain\% matchVec
}
\description{
An operator that matches the patterns on a vector of strings.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
