% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.interger64.R
\name{is.integer64}
\alias{is.integer64}
\title{Interger 64 Vectors
Evaluates if the elements on a vector are of class "interger64". Relevant for data.table functions.}
\usage{
is.integer64(x)
}
\arguments{
\item{x}{object to be tested.}
}
\value{
a boolean vector type TRUE if interger64, FALSE otherwise.
}
\description{
is.interger64 is a more general test of an object being interpretable as integers 64 bit encoded
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
