% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_cols.R
\name{keep_cols}
\alias{keep_cols}
\title{Keep Columns}
\usage{
keep_cols(DT, pattern, ...)
}
\arguments{
\item{DT}{a data.table to obtain column names from.}

\item{pattern}{a string containing the patterns to look for in the column names}

\item{...}{other arguments coming from the grep function}
}
\value{
a vector to be plugged into a data.table's "..j" for subsetting.
}
\description{
A function that emulates dplyr's .contains helper function combined with select, to subset columns combining a list and a pattern.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
