% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_perc.R
\name{missing_perc}
\alias{missing_perc}
\title{Percentage of Missing Observations}
\usage{
missing_perc(DT)
}
\arguments{
\item{DT}{a data.table to obtain column names from.}
}
\value{
a one row data.table keeping original names.
}
\description{
Calculates the percentage of missing observations for each column. It can be numeric (0.x) or presentation-ready (x\%).
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
