% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_summary.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Numeric Summaries}
\usage{
numeric_summary(x, na.rm)
}
\arguments{
\item{x}{a numeric vector containing the values to summarize.}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
a 1 x 4 data.table with summary statistics \code{data.table} including columns:
\itemize{
\item min
\item median
\item sd
\item max
}
}
\description{
Summarises numeric data and returns a data frame containing the minimum value, median, standard deviation, and maximum value.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
