% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_name.R
\name{object_name}
\alias{object_name}
\title{Get Object Name in character string}
\usage{
object_name(object)
}
\arguments{
\item{DT}{a data.table to obtain column names from.}
}
\value{
a character value or vector.
}
\description{
Retrieves the object name in a string character.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
