% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poltly_innteam.R
\name{plotly_innteam}
\alias{plotly_innteam}
\title{Innovation Team Plotly Theme}
\usage{
plotly_innteam(
  plot,
  ...,
  plot_title = "Insert Plot's Title",
  plot_subtitle = "",
  plot_caption = "",
  plot_xaxis = "",
  plot_yaxis = ""
)
}
\arguments{
\item{plot}{plotly plot}

\item{plot_title}{plot margin}

\item{plot_subtitle}{plot background}

\item{plot_caption}{font color}

\item{plot_xaxis}{font color}

\item{plot_yaxis}{font color}
}
\description{
Innovation Team Plotly Theme
}
\examples{
df <- data.table(State = state.name,
                LifeExp = state.x77[,4],
                Population = state.x77[,1])

df \%>\%
 plot_ly(x = ~LifeExp, y = ~Population, type = "scatter", width = 900, height = 500) \%>\%
   plotly_innteam(plot_title = 'This is my title',
                  plot_subtitle = 'Here we will see a subtitle',
                  plot_caption = 'Source: innteamUtils package',
                  plot_xaxis = 'LifeExp',
                  plot_yaxis = 'Population')
  )
}
