% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounder.R
\name{rounder}
\alias{rounder}
\title{Round Value}
\usage{
rounder(x, roundTo, dir = "up")
}
\arguments{
\item{x}{is a numeric vector or variable.}

\item{roundTo}{a numeric value indicating the closest number to round to.}

\item{dir}{a character value indicating wheter to round "up" or "down"}
}
\value{
an equal length vector.
}
\description{
It rounds bigger than 1 values, like to the closest 5, 10, 1000, etc
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
