% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_names.R
\name{snake_names}
\alias{snake_names}
\title{Snake Case Variable (column) names}
\usage{
snake_names(dat, ...)
}
\value{
a data.table
}
\description{
It modifies all column names into snake_case for internal practices consistency
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
