% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_right.R
\name{str_right}
\alias{str_right}
\title{Excel's RIGHT}
\usage{
str_right(string, i = 1, n)
}
\arguments{
\item{string}{string or vector containig strings to parse.}

\item{i}{character number from which to start counting.}

\item{n}{number of characters to extract.}
}
\value{
a vector with the remaining characters from the original strings
}
\description{
A function that emulates Excel's string functions LEFT(), MID() and RIGHT().
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
