% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_innteam.R
\name{theme_innteam}
\alias{theme_innteam}
\title{Innovation Team ggpolt2 Theme}
\usage{
theme_innteam(
  base_size = 12,
  base_family = "open-sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Contains the parameters and colour for a standarized plot theme.
}
\examples{
library(innteamUtils)
mtcars <- as.data.table(mtcars)
mtcars[, cyl := as.factor(cyl)] \%>\%
 ggplot(aes(x = mpg,
            y = hp,
            color = cyl)) +
 geom_point(size = 2, shape = 16) +
     scale_color_brewer(type = 'qual', palette = 7) +
     scale_fill_brewer(type = 'qual', palette = 7) +
         labs(title = 'Sample plot mtcars',
              subtitle = 'Another plot with mtcars but innteam theme',
              caption = 'The plot theme colour and fonts are identical with the rmarkdown file templates') +
 theme_innteam()

}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
