% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_q.R
\name{trim_q}
\alias{trim_q}
\title{Trim Variable}
\usage{
trim_q(x, lowbound = 0, upbound = 1)
}
\arguments{
\item{x}{is a numeric vector or variable.}

\item{lowbound}{lower bound percentile (in decimals).}

\item{upbound}{upper bound percentile (in decimals).}
}
\value{
a shorter or equal length vector.
}
\description{
It allows to cut tails simultaneously or single-sided (lower bound, upper bound). Quantiles range from 0 to 1.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
