% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_innteam.R
\name{theme_innteam}
\alias{theme_innteam}
\title{Innovation Team ggpolt2 Theme}
\usage{
theme_innteam(
  base_size = 12,
  base_family = "open-sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Contains the parameters and colour for a standarized plot theme.
}
\author{
Alejandro Abraham \href{mailto:a.abraham@innovationteam.eu}{a.abraham@innovationteam.eu}
}
